chore: RockPaperScissors파일 생성, InputError파일 생성
