+++ RockPaperScissors/RockPaperScissors/InputError.swift
@@ -1,0 +1,10 @@
+//
+//  InputError.swift
+//  RockPaperScissors
+//
+//  Created by 강민수 on 2022/12/26.
+//
+
+enum InputError: String, Error {
+    case invalidInput = "잘못된 입력입니다. 다시 시도해주세요"
+}
